requirements.txt
# ============================================
# DEPENDANCES CORE MLOps - Spam Detection
# ============================================

# Data Processing
pandas>=1.5.0
numpy>=1.23.0

# Machine Learning
scikit-learn>=1.2.0
joblib>=1.2.0

# Text Processing
nltk>=3.8.0
regex>=2023.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# ML Tracking
mlflow>=2.3.0
mlflow-skinny>=2.3.0  # Version légère

# Data Versioning
dvc>=3.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0
colorama>=0.4.6

# Web & Requests
requests>=2.28.0

# ============================================
# OPTIONNEL - Pour plus tard
# ============================================

# Additional Models
# xgboost>=1.7.0
# lightgbm>=4.0.0
# catboost>=1.2.0

# API
# fastapi>=0.100.0
# uvicorn[standard]>=0.23.0
# pydantic>=2.0.0

# Tests
# pytest>=7.3.0
# pytest-cov>=4.0.0